Ensure nova migration package is installed during minor update
Package management in puppet is disabled by default:
https://github.com/openstack/puppet-tripleo/blob/stable/newton/manifests/packages.pp

For a minor update we re-enable package management by setting enable_updates=true,
but only when there are updated packages available.

If the initial attempt to update fails it could have installed the updated packages
but not run the puppet manifest that install openstack-nova-migration. Subsequent
attempts to update will not resolve the issue as there are no longer updates available.

Change-Id: I4ea4b8def009ad11d8da7b82c365144fdd709e1e