Start using nic-configs from THT repo instead of tripleo-ci
This patch uses the net-iso gen tool to create multi-nic network
isolation templates with the same settings we currently use in CI.
Rather than overwrite existing example configs in the network dir,
we instead create a new dir under ci/environments to hold the
nic-configs we use in CI jobs. This alleviates any concerns about
backwards-compatibility and breaking users on upgrade.

We may want to also consider a massive cleanup of the existing
network dir with a proper deprecation period, but that is an
exercise left for later.

By moving these nic-configs to THT, we are bringing them inline
with the other heat environments we have moved out of the CI repo.
It also allows an end user to start from a configuration that we
have actually CI'd (though there are some settings like MTU that
are not actually good for a real deployment).

Change-Id: I9fd8b8bbcf18d26b1e05d33e2598c0d146814aff