Install ansible-pacemaker on O->P upgrade
This has been omitted in Ocata image, so we need to install this in
O->P upgrade again. Change Ic8ce72133c47a4c90d581a0925213877b11a471e
adds it to the image, so that we can stop installing it on P->Q
upgrade, and start using it for minor updates too.

Change-Id: I893792e8d82c716b2f3e7b8878b25ba89ea724f1
Co-Authored-By: Marius Cornea <mcornea@redhat.com>
Partial-Bug: #1701208