[newton only] Ensure cinder services are cleanup during upgrade.
It may happen that some leftover services definition stay in the
database.  We don't want to prevent upgrade because of that so we just
clean them up before going on.

Cleaning up the database while cinder services are running is not a
problem as the entry are recreated.  But duplicate and leftover /will/
be erased.

Closes-Bug: #1701259

Change-Id: I4ea3138e45aa430a89aad305fbb4b5eeefdfa16e