Disable network validation in multinode jobs
Sometimes the infracloud gateway refuses to ping even though
everything else is working fine.  Since we have coverage of this
functionality in the OVB jobs it should be safe to turn it off
here so it stops spuriously failing our jobs.

We can't just set the resource to OS::Heat::None because there
are other resources with dependencies on it.  Instead, this adds
a noop version of the validation software config that always
returns true.

Change-Id: I8361bc8be442b45c3ef6bdccdc53598fcb1d9540
Partial-Bug: 1680167