Use conditional to set heat workers
Heat is spawning number of workers equal to number of cores. This commit
fixes the problem by using a conditional so that os_workers fact is used
when the worker value is unset. This was changed in master in an
unrelated commit
https://github.com/openstack/tripleo-heat-templates/commit/aedb22be76be53716e30b60c93d323a3fbeb8f00
However, the above commit cannot be cherry picked.

Change-Id: I30b4ce6650c7f90d4a6bc96992baa201ce413fb8
Closes-Bug: #1702761