Corrected wrong os_region_name in glance-api conf file
KeystoneRegion value for all endpoints is set as 'regionOne',
it should be same in the configuration file.
In case of Cinder as glance backend the os_region_name should be
"regionOne" instead of "RegionOne".

Currently CI is not failing because cinder backend scenario is not yet added.
But this would definitely fail if os_region_name=RegionOne.

Change-Id: I26811a404a20ea3c55f5b272f86d9269d0f6acec
Closes-Bug: 1704060