Use static environment for deployed-server neutron mappings
We can't include these mappings for the deployed-server neutron ports
when using split-stack in the generated environment file because the
path to the actual templates directory is not known from with the
templates themselves.

This patch removes the mappings from the generated environment from the
stack output and adds a static environment file with relative paths that
will have to be included on the deploy command cli instead.

Change-Id: Id4b8c939fa7b26205609819b66e76bf73c9890d0
Closes-Bug: #1705144