Increase default RabbitMQ/Erlang TCP timeout from 5 to 15 seconds
This should be greater than the default value of
corosync_token_timeout, which is 10 seconds.  That way, if an entire
cluster node is unavailable, appropriate fencing measures can occur.

With the current settings, it is possible for brief network
interruptions, greater than 5 seconds, but less than 10 seconds, to
occur.  This can cause the RabbitMQ cluster to fail in subtle ways,
but no corrective action taken by pacemaker.

Change-Id: I735d43616c5c623c4398d924713012f595b2e5f9