nova_api_cron docker fix: add /var/spool/cron/nova
This patch reworks the nova_api_cron container so that it contains
the /var/spool/cron cron for the nova user and also so that it
contains the correct nova.conf file.

This should allow kolla-start to copy the correct config files
into place and then start the cron service to run the nova
tasks periodically.

Change-Id: Ib6b2ca5af5419130fb9c83f83d6f4bf97410e870
Related-bug: #1701254