Consistent hostname format env for split-stack
Adds a new environment split-stack-consistent-hostname-format.j2.yaml
for use with split-stack that will set the {{role.name}}HostnameFormat
parameters based on the actual roles in use.

It's possible to generate these parameters and values as long as the
roles files is the same between the 2 stacks, which will be documented
in I9997bd685f8f79537dfc377f9f1eb0c7446c3289.

Removes the hardcoded paramters from the other 2 split-stack
environments.

Change-Id: I93bc7c1a24a9cb2f9930372970c67c0eadb14add