Don't hardcode eth1 for undercloud nic config
eth1 will be the default as NeutronPublicInterface defaults to nic2,
which will be mapped to the second interface, which will be eth1 (all by
default).

So, there's no reason to hardcode to eth1 in the nic config template
directly, it can just use the special "interface_name" string which will
be replaced by the chosen nic at deployment time.

Change-Id: I97d1bb31cd8e24ed12ce916774aa83a6d11d410b