Use number for KeystoneCronTokenFlushMaxDelay instead of string
Using a string results in an erroneous check in puppet-keystone, which
sets up a zero where it shouldn't. So we change it to number to avoid
that. Note that there will also be a puppet-keystone fix for this.
Changing the value here assures that deployers only give valid values to
this parameter.

Change-Id: I00823e23358df91ce54f421c12636f05d4196e15
Closes-Bug: #1708584
(cherry picked from commit 79aca264fff6126edd98d480c2233cbddac64432)