teach yaml-validate.py how to be --quiet
This adds the --quiet (-q) option to yaml-validate.py.  With '-q',
yaml-validate.py will only print warnings or errors. With '-qq',
yaml-validate.py will print only errors. This makes it much easier to
spot problems when working on templates.

This commit does not change the default behavior of yaml-validate.py.

Change-Id: I358f1d4edde03714627d98361b44e6b90ce5e93c