Correctly set DEFAULT/dhcp_domain to empty string.
Only for stable/newton the DEFAULT/dhcp_domain setting in nova.conf is
not set correctly to the empty sring, but not set at all, falling back
to the default value. Later branches and master don't show this
behavior so there is no need for a fix there. This possibly is caused by
a difference between puppet 3 vs. puppet 4.


Change-Id: I5bdbef6bd170f9f2dae4fd5e74e478a19a02e755
Closes-Bug: #1449520