[newton only] Cinder services cleanup should happen when mysql is up
In a previous patch we attempt to fix a issue with old entries in the
cinder service database making the upgrade fail[1].  The fix was wrong
as the cinder-manage command was triggered after the galera shutdown,
making it useless.

This patch re-order the pacemaker shutdown process to have a state
where cinder-volume is stopped and pacemaker is up so that we can run
cinder-manage service cleanup safely.

[1] https://review.openstack.org/#/c/478922/

Closes-Bug: #1701259

Change-Id: I4ee6799b32820a681ffcbf339effdf86337485f7