Remove External network from Networker role NIC configs
This change removes the External network from the Networker
role, since it is not used or needed on that role.

The External network appers in the NIC config templates for the
Networker role, but this network is not used by the Networker
role. This results in deployment errors, since no IP address
is created for the External network, and the Networker roles
cannot reach an external gateway.

Change-Id: I78e0c9b50a7fee8efeea1ab639b44e0c6f2aa922
Closes-bug: 1720257