Remove Heat Cloudwatch API during upgrade and disable by default
This adds a heat-api-cloudwatch-disabled.yaml and wires it up in
the resource registry. During the Ocata to Pike upgrade this service
will thus be stopped and disabled by default.

If you wish to keep the Heat Cloudwatch API then you should instead
use the provided heat-api-cloudwatch.yaml environment file.

Change-Id: I3f90a9799b90ca365f675f593371c1d3701fede6
Related-Bug: 1713531