our undercloud default nic should be eth1
In commit 2ffc62209bf14b7eb1c8191a8ef65e6b037c9604 (Don't hardcode eth1
for undercloud nic config) we removed the hard coded nic and instead
rely on the default value from the NeutronPublicInterface heat parameter.
As NeutronPublicInterface defaults to nic1 in overcloud.j2.yaml this
functionally broke the undercloud defaults as we typically expect eth1
to be the bridged interface used for the control_plane network traffic.

This patch sets the correct default again by setting the value
manually in the undercloud heat environment.

NOTE: using eth1 per the old hard coded value. Eventually we want
this to be nic2 but there has been a regression in os-net-config
which is causing "active nics" not to be detected correctly in
all cases.

Change-Id: Idb1e34bd3dff7c6412082ab4201f8e6af35a939d
Closes-bug: #1721853