Revert "Fixes heat resource name for Internal API Network"
This reverts commit 520be6bb4056ead8e6fad08ad96e99f7da5b341e.

This introduced a bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1501515

where during upgrade, the previous heat resource would for the
InternalApi network would have the incorrect name "Internal" and the
upgrade would try to delete the resource in order to create
"InternalApi".  This needs to be reverted and a proper fix will be
submitted that accounts for this upgrade scenario.

Related-Bug: #1718764

Change-Id: Id906fac421db317ce48d5cecfcd43397a0f4ab3d