Revert "Fixes heat resource name for Internal API Network"
This reverts commit 97244b942d29d2b5acd7a3eb07acdba0d9b99677.

This introduced a bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1501515

where during upgrade, the previous heat resource would for the
InternalApi network would have the incorrect name "Internal" and the
upgrade would try to delete the resource in order to create
"InternalApi".  This needs to be reverted and a proper fix will be
submitted that accounts for this upgrade scenario.

Related-Bug: #1718764

Change-Id: Ied908020ed856a5573f1333b9139029d0ffc37b4