Start using ipv6 nic-configs from THT repo instead of tripleo-ci
This patch uses the net-iso gen tool to create ipv6 multi-nic network
isolation templates with the same settings we currently use in CI.
Rather than overwrite existing example configs in the network dir,
we instead create a new dir under ci/environments to hold the
nic-configs we use in CI jobs. This alleviates any concerns about
backwards-compatibility and breaking users on upgrade.

By moving these nic-configs to THT, we are bringing them inline
with the other heat environments we have moved out of the CI repo.
It also allows an end user to start from a configuration that we
have actually CI'd.

Change-Id: Iba56667e11510a1cc2470a7dfdc6bc0158878010