Disable SwiftDispersion when using docker
We currently have the following in the registry:
OS::TripleO::Services::SwiftDispersion: puppet/services/swift-dispersion.yaml

Since this service is included by default in the Controller role
it will be installed on the host even on a containerized deployment.

Let's noop this in docker.yaml until a containerized version of it
gets merged.

Change-Id: Ic2793d0cfb7b20f4661cb1a45793cae67a4868b4
Closes-Bug: #1723788