Allow installing multiple-node Kubernetes
By setting loadbalancer_apiserver_localhost to false we tell the
kubelets to register with 1st master rather than assuming there's an
API proxy running on each host.

Also Kubespray expects a specific format of inventory because it tries
to enforce that hostname matches the invenotry name of the node. This
previously resulted in incorrect hostnames being set.

And we also open the necessary firewall ports to allow the cluster to
communicate.

Change-Id: I1c0251be58535d59fe0ec157cb9acaba209e8732