Add NetIpMap to hieradata for *ExtraConfig overrides
To enable per-node override of bind IPs via the per-role
ExtraConfig paramaters, we need to enable hiera interpolation
that references the keys defined in NetIpMap, so we add them
to the hieradata.  To minimise the risk of any conflicts in
keynames it's added near the bottom of the hierarchy, but
I'm not aware of any conflicting names in our templates/modules.

This will allow per-node hieradata override of bind IPs e.g:

  parameter_defaults:
    ComputeRack1ExtraConfig:
      nova::vncproxy::host: "%{hiera('rack1_internal_api')}"
    ComputeRack2ExtraConfig:
      nova::vncproxy::host: "%{hiera('rack2_internal_api')}"

Closes-Bug: #1726884
Change-Id: Icf7da1d78176c2ee0197ff2459d69d995cbb16ad