os-net-config not run when conf is empty and re-introduce exception.
There may be cases where os-net-config is empty leading to issue when
we force the run during its special treatment in upgrade/update.

Piggy back in this review is re-adding the os-net-config special
handling to update and non-controller.  But it’s more robust now it
won’t run the command unconditionally, it needs:
 - os-net-config to be upgraded
 - /etc/os-net-config/config.json to exist and not be empty.

Another improvement is that it will prevent any networking cut during
its run as we are using the –no-activate flag.

Change-Id: Ie1b88d5cac22fca5b0b325b1dd8c81d65b63acb8
Closes-Bug: #1728889