Update ansible before the ansible upgrade tasks
In case an ansible update is available during the upgrade then
the ansible package gets updated to a new version by the ansible
tasks. This could potentially lead to issues as the one described
in LP#1729546. This change updates the ansible package via yum
before starting the ansible upgrade tasks in order to avoid having
ansible updating itself.

Related-bug: 1729546

Change-Id: I2ea0aa1f670053578996018663c9fa52dec14b77