Create Swift directory d1 if required
Swift added a requirement that storage directories must exist before
using them. In case of the d1 directory in TripleO - used when there are
no "real disks" - it has to be created by TripleO in advance.

This change forwards the SwiftUseLocalDir parameter to puppet-tripleo,
which creates the directory with the correct permissions.

Closes-Bug: 1729569
Depends-On: I49e395ac379ced01adb60d8d9f951c08718b1c61
Change-Id: I82b783541c6097f6b8747b63349720c47f3b3f94