Fix rights to local_settings.d for dockerized Horizon
For some reasonf that directory doesn't have r/x rights, so when
compress is ran as root, it can access config files in it, but when
horizon is run by apache, it can't, and expects different theme files,
thus failing with OfflineGenerationError. Giving apache access to that
directory fixes the problem and makes the custom theme work.

Closes-bug: #1730911
Change-Id: I53f6db23b036bc9b5a689bbac958550f384194c6