OVN: Provide the option to define NeutronBridgeMappings as a role parameter
In order for a node to provide external connectivity to tenant traffic,
ovn-bridge-mappings should be defined which is read by ovn-controller service.
This value comes from the T-H-T param 'NeutronBridgeMappings'. Presently
this param is defined with the default value "datacentre:br-ex" and is configured
on every node where ovn-controller service is enabled. This would cause problems
if some of the hosts do not provide external connectvity.

To solve this problem, this patch adds 'NeutronBridgeMappings' as a role parameter
so that it can set to empty string if the role doesn't provide external connectivity.

In the OVN environment template files, this param is set to empty for Compute role.

Since NeutronBridgeMappings is also used by neutron-ovs-agent service, this patch
also modifies puppet/services/neutron-ovs-agent.yaml to define NeutronBridgeMappings
as a role parameter.

Change-Id: I6a00b8dc1ff387cc5e1e62b4d9d7da5e1239ee31
Closes-bug: #1730711